### 2019-01-20 #1178 @labarba
@mtezzele — I wonder if you'd rather use the word "morphing" instead of "deformation" in the title. When I read "deformation" I think of the elastic or plastic deformation of engineered objects under stress.
