### 2019-01-20 #1178 @mtezzele
@labarba thank you very much for your suggestions. I will do the modifications and ask few people if they are available as potential reviewers. I will let you know ASAP.
