+++ joss-paper/paper.bib
@@ -17,5 +17,5 @@
 	Year = {2018}}
 
 @url{bladex,
-	Title = {{BladeX: Python Blade Deformation. Available at: \url{https://github.com/mathLab/BladeX}}},
+	Title = {{BladeX: Python Blade Morphing. Available at: \url{https://github.com/mathLab/BladeX}}},
 	Urldate = {2019-01}}
