### 2019-01-21 #1178 @labarba
Thank you, @mtezzele — that helps a lot.

👋  @benoitpaillard, @spiani — would you be willing to sign-up as reviewers for JOSS?
http://joss.theoj.org/reviewer-signup.html
That will give me more information about your areas of expertise (but also leaves you open for future review invitations! ... which you can always decline, of course).
If not, you could email me with your real name and a few words about your expertise. Thanks! 🙏
Do take a look at the [JOSS Review Process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
