+++ .travis.yml
@@ -9,16 +9,16 @@ matrix:
     include:
         - os: linux
           python: 2.7
-          env: TOXENV=py27
+          env: TOXENV=py27 CONDA_PY=27
         - os: linux
           python: 3.6
-          env: TOXENV=py36
+          env: TOXENV=py36 CONDA_PY=36
         - os: osx
           language: generic
-          env: TOXENV=py27
+          env: TOXENV=py27 CONDA_PY=27
         - os: osx
           language: generic
-          env: TOXENV=py36
+          env: TOXENV=py36 CONDA_PY=36
           
 before_script:
     - "export DISPLAY=:99.0"
@@ -30,20 +30,13 @@ before_script:
 before_install:
     # We do this conditionally because it saves us some downloading if the
     # version is the same.
-    - if [[ "$TRAVIS_PYTHON_VERSION" == "2.7" ]]; then
+    - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then
         wget https://repo.continuum.io/miniconda/Miniconda2-latest-Linux-x86_64.sh -O miniconda.sh;
-      fi
-    - if [[ "$TRAVIS_PYTHON_VERSION" == "3.6" ]]; then
-        wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
       fi
     - if [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
         brew update;
         brew install python;
-        if [[ "$TOXENV" == "py27" ]]; then
-            wget https://repo.continuum.io/miniconda/Miniconda2-latest-MacOSX-x86_64.sh -O miniconda.sh;
-        else
-            wget https://repo.continuum.io/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O miniconda.sh;
-        fi
+        wget https://repo.continuum.io/miniconda/Miniconda2-latest-MacOSX-x86_64.sh -O miniconda.sh;
       fi
     - python --version
     - chmod +x miniconda.sh
@@ -67,6 +60,7 @@ install:
         fi
       fi
     - source activate test
+    - python --version
     - conda install --yes numpy scipy matplotlib pip nose sip=4.18
     - if [[ "$TOXENV" == "py27" ]]; then
         conda install --yes -c conda-forge -c dlr-sc -c pythonocc -c oce pythonocc-core=0.18.1 smesh=6.7.5 python=2.7;
@@ -76,7 +70,6 @@ install:
     - pip install setuptools
     - pip install enum34
     - pip install coveralls
-    - pip install coverage
     - python setup.py install
 
 script:
