### 2019-01-21 #1178 @benoitpaillard
Hi, I just signed up to JOSS but got no email back. Let me know if that is normal.
