### 2019-01-21 1f257e4
removed osx with py36 because of coveralls issue
