### 2019-01-23 #1178 @spiani
I have registered on JOSS, but I also have received no feedback. Could you please add me as a reviewer? Thank you
