### 2019-01-23 #1178 @arfon
> I have registered on JOSS, but I also have received no feedback. Could you please add me as a reviewer? Thank you

We have you as a potential reviewer in our spreadsheet now (although I note you didn't specify what are your areas if interest/specialist topics). I'm not sure if @labarba is planning on asking for your help reviewing this submission - @labarba?
