### 2019-01-25 #1178 @spiani
@labarba For sure. Where can I send you a brief description of my background?
