### 2019-01-26 #1203 @whedon
[REVIEW]: BladeX: Python Blade Deformation
**Submitting author:** @mtezzele (<a href="http://orcid.org/0000-0001-9747-6328">Marco Tezzele</a>)
**Repository:** <a href="https://github.com/mathLab/BladeX" target ="_blank">https://github.com/mathLab/BladeX</a>
**Version:** v0.1
