### 2019-01-26 #1203 @labarba
@benoitpaillard, @Juanlu001 — this is where the action happens. You can read about the [JOSS Review Process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) in our docs. Each of you has a review checklist at the top of this issue 👆 and you should feel free to ask questions here, and open issues in the submission repository as needed.
