### 2019-01-27 #1203 @labarba
@mtezzele — I notice that you cited the software repository in your paper. The paper format will have a link to the repository (and software archive, made after review is complete) on the front page, so it's not necessary to add it to the References list. You also cite a _submitted_ paper: would you be able to post a pre-print of the manuscript, for example in [engrXiv](https://engrxiv.org) or a similar service? (Then you can add the DOI or pre-print identifier in your References list.)
