### 2019-01-27 mathLab/BladeX#42 @Juanlu001
Suggestion: modernize installation instructions
A workaround is shown in the installation instructions to be able to uninstall the package:

https://github.com/mathLab/BladeX/blame/v0.1/README.md#L63-L68

Nowadays, the most common practice is to do `pip install git+https://github.com/mathLab/BladeX.git` or `pip install /path/to/BladeX` instead, which allows easy uninstall and is forward-compatible with [a future world without `setup.py` files](https://www.python.org/dev/peps/pep-0517/).
