### 2019-01-28 mathLab/BladeX#43 @Juanlu001
Suggestion: clarify support channels
In the "Authors and contributors" section of the README, users are encouraged to contact @mahgadalla and @mtezzele by email. Although it's crystal clear, I would suggest to include a explicit "Support" section in the README stating that, and also clarifying whether GitHub issues are welcomed.
