### 2019-01-28 #1203 @Juanlu001
I did a first pass on the repository and left two minor suggestions, https://github.com/mathLab/BladeX/issues/42 and https://github.com/mathLab/BladeX/issues/43 (none of them a blocker). I will finish my review during the coming days.
