### 2019-01-28 a7f664c
add arxiv, remove unnecessary ref
