+++ joss-paper/paper.bib
@@ -13,9 +13,6 @@
 	Author = {Tezzele, M. and Demo, N. and Mola, A. and Rozza, G.},
 	Preprint = {https://arxiv.org/abs/1810.12364},
 	Journal = {Submitted, Special Volume ECMI},
+	Preprint = {https://arxiv.org/abs/1810.12364},
 	Title = {{An integrated data-driven computational pipeline with model order reduction for industrial and applied mathematics}},
 	Year = {2018}}
-
-@url{bladex,
-	Title = {{BladeX: Python Blade Morphing. Available at: \url{https://github.com/mathLab/BladeX}}},
-	Urldate = {2019-01}}
