### 2019-01-28 #1203 @mtezzele
@labarba I added the Preprint record to the bibitem but it does not appear in the pdf (https://github.com/mathLab/BladeX/pull/45). Can you please tell me how to do it?
