### 2019-01-28 #1203 @labarba
@mtezzele — I believe you should use the `note` field to provide the URL to the preprint. Also, the entry type should probably be `@unpublished` or `@misc`.
