### 2019-01-28 #1203 @labarba
Dang! Try with the `howpublished` field in `@misc`.
