+++ joss-paper/paper.bib
@@ -12,6 +12,6 @@
 @misc{tezzele2018ecmi,
     author = {Tezzele, M. and Demo, N. and Mola, A. and Rozza, G.},
     journal = {Submitted, Special Volume ECMI},
-    note = {https://arxiv.org/abs/1810.12364},
+    howpublished = {https://arxiv.org/abs/1810.12364},
     title = {{An integrated data-driven computational pipeline with model order reduction for industrial and applied mathematics}},
     year = {2018}}
