### 2019-02-03 #1203 @benoitpaillard
Hi all, sorry for the late completion.

The really only issue I'd mention is the dependency to pythonocc which is still a mess to compile, but don't get me wrong, I'm using it in my projects as well ! I just wish it could be compiled more easily.

Appart from that, finally a code that does 3D blades, but also robust naca and general airfoil generators !! Congrats guys, great tool.
