### 2019-02-03 #1203 @labarba
@mtezzele — Some final editorial comments:

¶1.
obtain designs which allow >> that allow
prototipation environment >> prototyping environment
By a practical standpoint, the propeller blade >> From a practical standpoint, a propeller blade
the shape of a certain number of sectional profile shapes >> the shape of a certain number of sectional profiles
a morphing tool that expand >> expands
In particular we followed >> In particular, we followed [add comma]

¶2.
In the BladeX Python package we implemented >> package, we… [add comma]
simulation driven design >> simulation-driven design [add hyphen]

¶3.
naval propulsion”, >>naval propulsion,”  [comma inside quote]
can be found in (Tezzele, Demo, Mola, & Rozza, 2018) >> this is an in-text citation (meaning, the citation is part of the sentence), so you need to suppress the parenthesis; see [Citation Syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)

¶4.
the resulted iges and stl  le formats >> the resulting
(same in caption of Figure 1)
