### 2019-02-04 01562a2
joss editorial fixes
