### 2019-02-04 #1203 @mahgadalla
@labarba thanks a lot! the editorial comments are now fixed. Please let us know if the in-text citation looks ok now, also if the first citation should be regarded as in-text as well?
