### 2019-02-04 #1203 @labarba
Yes, I missed that one. Whenever the citation is part of the sentence, treat as in-text.
